implement important functionality for entities and actions
Trying to walk into a tile now uses the action `Bump`, which will in turn call either `Move` or `Melee` as needed. When the player performs an action, other entities get a "turn", though it doesn't do much right now.
`Melee` will create an attack interaction between the performing entity and the entity at the target position.

Some of the code has been restructured and separated as well.