implement player inventory and a health potion
The inventory has a max size of 26 and when using/dropping an item there's a conversion from abc to a numeric index. Currently only a health potion is implemented. The effect is implemented in several parts: A Healing component storing information about an item that should be able to heal, and an IsQuaffable tag which tells the engine the potion can be quaffed.